+++ NEWS.md
@@ -1,3 +1,5 @@
+# labelled (development version)
+
 # labelled 2.7.0
 
 * a `.values` argument has been added to `set_na_values()` and `set_na_range()`, allowing to pass a list of values