+++ NEWS.md
@@ -1,4 +1,4 @@
-# labelled (development version)
+# labelled 2.8.0
 
 * new helper `recode_if()` for recoding values based on condition, variable
   and value labels being preserved (#82)